msm8996-common: Import qti power HAL
* Rename it to android.hardware.power-service-x2
* Fix utils.c for our 4core msm8996
* Add power hal to file_contexts

* Fixing following errors while Enforcing:

I servicemanager: Since 'android.hardware.power.IPower/default' could not be found, trying to start it as a lazy AIDL service
04-08 09:19:56.602
W libc : Unable to set property "ctl.interface_start" to "aidl/android.hardware.power.IPower/default": error code: 0x20
04-08 09:19:56.603
I servicemanager: Tried to start aidl service android.hardware.power.IPower/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.