msm8996-common: wifi: Configure correct overlay configuration.
The vendor is working to support it in T. (b/183916225)
The fix corrects the overlay configuration, which introduced
for android T.

PS: The impact of the configuration is two cases.
1. Configuration should be false but it configures to true:
It will cause no CC(country code) changed event to notify listener.
It might cause any function fail which relies on CC.
(It is happening on P21 series)

2. Configuration should be true but it configures to false:
Minor impact than #1 because it just update CC before the Wifi subsystem
apply it.5. It might cause the data to be incorrect because of the
timing issue.

Bug: 183916225
Test: Manuel Test to check that configuration is correct.
Change-Id: Ica764e4a22def6dea0613882b3ae403852a03a83